pyproject.toml
[project]
name = "weatherapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-limiter>=0.1.6",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
